requirements.txt
torch
torchvision
matplotlib
seaborn
numpy
pandas
pillow
scikit-learn
grad-cam
streamlit
tqdm